Bridge the ALVR backend to the negotiated display-clock frame pacer
Add the OXRSys side of the ALVR frame-pacing extension: a versioned C API mirror
(alvr_set_frame_pacing_mode, capabilities/timesync/feedback events, paced video
send), an alvr_frame_pacing config knob (off/shadow/on), and AlvrStreamingBackend
wiring that owns a FramePacer, feeds it client display/timesync/feedback events,
maintains an unsettled-target ring, and only drives closed-loop release in active
mode. Session selects backend pacing without regressing the stock-client grid;
display-lag attribution is gated on fresh feedback so reuse cannot drift the servo.

+++ runtime/src/FramePacer.cpp
@@ -71,7 +71,8 @@ void FramePacer::SetNominalPeriod(int64_t periodNs)
     timelineLocked_ = false;
 }
 
-void FramePacer::OnClientTiming(uint32_t sessionEpoch,int64_t clientPredictedDisplayNs, int64_t clientPeriodNs)
+void FramePacer::OnClientTiming(uint64_t sessionEpoch, int64_t clientPredictedDisplayNs,
+                                int64_t clientPeriodNs)
 {
     if (clientPeriodNs <= 0)
     {
